# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/organizations": {
      "post": {
        "description": "Creates an organization for the subscription of the authenticated user. Subject to the plan limit max_organizations. The creator receives the Owner role with full permissions.",
        "operationId": "OrganizationController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganizationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Organization created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (one message per invalid field) or a business rule was violated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found, or it belongs to another organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited. The `default` bucket allows 300 requests per 60 s per client IP; the auth routes use the stricter `auth` bucket.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create an organization",
        "tags": [
          "organizations"
        ],
        "x-auth": {
          "required": true,
          "scheme": "bearer",
          "context": "subscription"
        },
        "x-organization-context": {
          "required": false
        }
      },
      "get": {
        "description": "Organizations owned by the subscription of the authenticated user (tenant isolation, spec §4.3).",
        "operationId": "OrganizationController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Page of organizations owned by the subscription",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationPage"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (one message per invalid field) or a business rule was violated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found, or it belongs to another organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited. The `default` bucket allows 300 requests per 60 s per client IP; the auth routes use the stricter `auth` bucket.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List organizations",
        "tags": [
          "organizations"
        ],
        "x-auth": {
          "required": true,
          "scheme": "bearer",
          "context": "subscription"
        },
        "x-organization-context": {
          "required": false
        }
      }
    },
    "/organizations/{organizationId}": {
      "get": {
        "description": "One organization of the subscription. The :organizationId path segment defines the active organization for this call; your membership in it is verified.",
        "operationId": "OrganizationController_findOne",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "description": "Active organization id. Optional on /organizations/:organizationId routes, required on every other organization-scoped route.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Organization details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (one message per invalid field) or a business rule was violated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Permission denied, or you are not a member of the active organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited. The `default` bucket allows 300 requests per 60 s per client IP; the auth routes use the stricter `auth` bucket.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get an organization",
        "tags": [
          "organizations"
        ],
        "x-auth": {
          "required": true,
          "scheme": "bearer",
          "context": "organization"
        },
        "x-organization-context": {
          "required": true,
          "header": "x-organization-id",
          "pathParameter": "organizationId"
        },
        "x-permissions": [
          "organization.view"
        ]
      },
      "patch": {
        "description": "Renames an organization or changes its lifecycle status. Only fields you send are modified.",
        "operationId": "OrganizationController_update",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "description": "Active organization id. Optional on /organizations/:organizationId routes, required on every other organization-scoped route.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrganizationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (one message per invalid field) or a business rule was violated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Permission denied, or you are not a member of the active organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited. The `default` bucket allows 300 requests per 60 s per client IP; the auth routes use the stricter `auth` bucket.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update an organization",
        "tags": [
          "organizations"
        ],
        "x-auth": {
          "required": true,
          "scheme": "bearer",
          "context": "organization"
        },
        "x-organization-context": {
          "required": true,
          "header": "x-organization-id",
          "pathParameter": "organizationId"
        },
        "x-permissions": [
          "organization.update"
        ]
      },
      "delete": {
        "description": "Deletes the organization and every record scoped to it (roles, teams, clients, projects, tasks, time), including anything already soft-deleted. Unlike the other delete routes this is permanent, so it requires the organization name to be typed as ?confirm=<name>; without it the request is refused with 409 and nothing is deleted.",
        "operationId": "OrganizationController_remove",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "confirm",
            "required": false,
            "in": "query",
            "description": "The organization name, typed exactly, to confirm permanent deletion. Anything else is refused with 409 and nothing is deleted.",
            "schema": {
              "example": "Acme Field Services",
              "type": "string"
            }
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "description": "Active organization id. Optional on /organizations/:organizationId routes, required on every other organization-scoped route.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Organization deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationMessageResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (one message per invalid field) or a business rule was violated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Permission denied, or you are not a member of the active organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflicting state: the resource already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited. The `default` bucket allows 300 requests per 60 s per client IP; the auth routes use the stricter `auth` bucket.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete an organization",
        "tags": [
          "organizations"
        ],
        "x-auth": {
          "required": true,
          "scheme": "bearer",
          "context": "organization"
        },
        "x-organization-context": {
          "required": true,
          "header": "x-organization-id",
          "pathParameter": "organizationId"
        },
        "x-permissions": [
          "organization.delete"
        ]
      }
    }
  },
  "info": {
    "title": "OPS API",
...
    "version": "1.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400,
            "description": "HTTP status code"
          },
          "message": {
            "oneOf": [
              {
                "type": "string",
                "example": "Validation failed"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": [
                  "name should not be empty",
                  "email must be an email"
                ]
              }
            ],
            "description": "Human readable summary: one string, or one string per invalid field"
          },
          "error": {
            "type": "string",
            "enum": [
              "Bad Request",
              "Unauthorized",
              "Forbidden",
              "Not Found",
              "Conflict",
              "Too Many Requests"
            ],
            "example": "Bad Request",
            "description": "Status name. Absent on the 401 of a rejected or expired JWT and on every 429, because those exceptions are thrown without a message."
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
...
      "CreateOrganizationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Acme Delivery",
            "description": "Organization name"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "archived"
            ],
            "default": "active",
            "description": "Lifecycle status"
          }
        },
        "required": [
          "name"
        ]
      },
      "OrganizationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "4c1f5c66-2f9e-4f5c-8b2a-9c0d1e2f3a4b"
          },
          "tenantId": {
            "type": "string",
            "example": "b7d8e9f0-1a2b-4c3d-8e4f-5a6b7c8d9e0f"
          },
          "name": {
            "type": "string",
            "example": "Acme Delivery"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "archived"
            ],
            "example": "active"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-09-30T10:00:00.000Z"
          }
        },
        "required": [
          "id",
          "tenantId",
          "name",
          "status",
          "createdAt"
        ]
      },
      "OrganizationPage": {
        "type": "object",
        "properties": {
          "items": {
            "description": "Records of the current page",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrganizationResponseDto"
            }
          },
          "total": {
            "type": "number",
            "example": 42,
            "description": "Total matching records"
          },
          "page": {
            "type": "number",
            "example": 1,
            "description": "Current page (1-based)"
          },
          "limit": {
            "type": "number",
            "example": 20,
            "description": "Items per page"
          },
          "totalPages": {
            "type": "number",
            "example": 3,
            "description": "Total number of pages"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "limit",
          "totalPages"
        ]
      },
      "UpdateOrganizationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Acme Delivery EU"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "archived"
            ]
          }
        }
      },
      "OrganizationMessageResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Organization deleted"
          }
        },
        "required": [
          "message"
        ]
      }
    }
  }
}